customize for itch & add audios

+++ app/js/prefabs/Player.js
@@ -32,30 +32,6 @@ YellowSidd.Player = function (game_state, name, position, properties) {
   this.shoot_timer = this.game_state.game.time.create();
   this.shoot_timer.loop(Phaser.Timer.SECOND / this.attack_rate, this.shoot, this);
 
-  // Add extra lives according to made extra heart purchases.
-  if (localStorage.heart_plus_one_once === '1') {
-    this.game_state.prefabs.player.lives += 1;
-  }
-  if (localStorage.heart_plus_one_infinite === '1') {
-    this.game_state.prefabs.player.lives += 1;
-  }
-
-  // Add extra speed according to made extra speed purchases.
-  if (localStorage.speed_plus_one_once === '1') {
-    this.game_state.prefabs.player.walking_speed *= 1.5;
-  }
-  if (localStorage.speed_plus_one_infinite === '1') {
-    this.game_state.prefabs.player.walking_speed *= 1.5;
-  }
-
-  // Add extra jump according to made jump heart purchases.
-  if (localStorage.jump_plus_one_once === '1') {
-    this.game_state.prefabs.player.jumping_speed *= 1.5;
-  }
-  if (localStorage.jump_plus_one_infinite === '1') {
-    this.game_state.prefabs.player.jumping_speed *= 1.5;
-  }
-
   // Add sounds.
   this.fireball_sound = this.game.add.audio('fireball');
   this.jump_sound = this.game.add.audio('jump');
@@ -82,13 +58,23 @@ YellowSidd.Player.prototype.update = function () {
     this.body.velocity.x = this.walking_speed;
     this.direction = "RIGHT";
     this.animations.play('walking'); // Play walking animation.
-    this.walking_sound.play(); // Play walking sound.
+
+    // Play sound only if player left button sound as on mode.
+    if (PLAY_SOUND) {
+      this.walking_sound.play(); // Play walking sound.
+    }
+
     this.scale.setTo(1, 1);
   } else if (this.cursors.left.isDown && this.body.velocity.x <= 0) {
     // Move left
     this.body.velocity.x = -this.walking_speed;
     this.direction = "LEFT";
-    this.walking_sound.play(); // Play walking sound.
+
+    // Play sound only if player left button sound as on mode.
+    if (PLAY_SOUND) {
+      this.walking_sound.play(); // Play walking sound.
+    }
+
     this.animations.play('walking'); // Play walking animation.
     this.scale.setTo(-1, 1);
   } else {
@@ -101,7 +87,11 @@ YellowSidd.Player.prototype.update = function () {
   // Jump only if touching a tile.
   if (this.cursors.up.isDown && this.body.blocked.down) {
     this.body.velocity.y = -this.jumping_speed;
-    this.jump_sound.play(); // Play jump sound.
+
+    // Play sound only if player left button sound as on mode.
+    if (PLAY_SOUND) {
+      this.jump_sound.play(); // Play jump sound.
+    }
   }
 
   // Dies if touches the end of the screen.
@@ -115,7 +105,11 @@ YellowSidd.Player.prototype.update = function () {
       // Start a timer which will call the shoot method in a loop.
       this.shoot();
       this.shoot_timer.start();
-      this.fireball_sound.play(); // Play jump sound.
+
+      // Play sound only if player left button sound as on mode.
+      if (PLAY_SOUND) {
+        this.fireball_sound.play(); // Play jump sound.
+      }
     }
   } else {
     this.shoot_timer.stop(false);
@@ -129,7 +123,11 @@ YellowSidd.Player.prototype.hit_enemy = function (player, enemy) {
     this.score += enemy.score; // Add points to score.
     enemy.kill(); // Kill enemy.
     player.y -= this.bouncing;
-    this.kill_enemy_sound.play(); // Play kill enemy sound.
+
+    // Play sound only if player left button sound as on mode.
+    if (PLAY_SOUND) {
+      this.kill_enemy_sound.play(); // Play kill enemy sound.
+    }
   } else {
     this.die(); // Restart level.
   }
@@ -138,7 +136,12 @@ YellowSidd.Player.prototype.hit_enemy = function (player, enemy) {
 YellowSidd.Player.prototype.die = function () {
   "use strict";
   this.lives -= 1;
-  this.lost_heart_sound.play(); // Play lost heart sound.
+
+  // Play sound only if player left button sound as on mode.
+  if (PLAY_SOUND) {
+    this.lost_heart_sound.play(); // Play lost heart sound.
+  }
+
   this.shooting = false;
 
   if (this.lives > 0) {
